+++ src/constants/appVersion.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = "13.139.9";
+export const APP_VERSION = "13.139.10";